Adds google analytics and google tag manager #1517
