+++ src/config.js
@@ -36,6 +36,23 @@ const editorialFrontendDomain = () => {
   }
 };
 
+const gaTrackingId = () => {
+  if (process.env.NODE_ENV !== 'production') {
+    return '';
+  }
+
+  switch (ndlaEnvironment) {
+    case 'local':
+      return '';
+    case 'dev':
+      return '';
+    case 'prod':
+      return 'UA-9036010-36';
+    default:
+      return 'UA-9036010-36';
+  }
+};
+
 const learningpathFrontendDomain = () => {
   switch (ndlaEnvironment) {
     case 'local':
@@ -102,6 +119,8 @@ const config = {
   enableFullTaxonomy: ndlaEnvironment === 'test' || ndlaEnvironment === 'local',
   localConverter: getEnvironmentVariabel('LOCAL_CONVERTER', false),
   checkArticleScript: getEnvironmentVariabel('CHECK_ARTICLE_SCRIPT', false),
+  googleTagManagerId: getEnvironmentVariabel('NDLA_GOOGLE_TAG_MANAGER_ID'),
+  gaTrackingId: gaTrackingId(),
 };
 
 export function getUniversalConfig() {
