+++ src/containers/LearningResourcePage/EditLearningResource.jsx
@@ -9,9 +9,9 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import Helmet from 'react-helmet';
 import { Redirect, withRouter } from 'react-router-dom';
-
+import { HelmetWithTracker } from '@ndla/tracker';
+import { injectT } from '@ndla/i18n';
 import * as messageActions from '../Messages/messagesActions';
 import { actions, getDraft } from '../../modules/draft/draft';
 
@@ -69,7 +69,7 @@ class EditLearningResource extends PureComponent {
   }
 
   render() {
-    const { article, selectedLanguage, ...rest } = this.props;
+    const { article, selectedLanguage, t, ...rest } = this.props;
     if (!article) {
       return null;
     }
@@ -85,7 +85,9 @@ class EditLearningResource extends PureComponent {
       : selectedLanguage;
     return (
       <div>
-        <Helmet title={`${article.title} - NDLA`} />
+        <HelmetWithTracker
+          title={`${article.title} ${t('htmlTitles.titleTemplate')}`}
+        />
         <LearningResourceForm
           initialModel={getInitialModel(article, language)}
           selectedLanguage={selectedLanguage}
@@ -136,9 +138,11 @@ const mapStateToProps = (state, props) => {
   };
 };
 
-export default withRouter(
-  connect(
-    mapStateToProps,
-    mapDispatchToProps,
-  )(EditLearningResource),
+export default injectT(
+  withRouter(
+    connect(
+      mapStateToProps,
+      mapDispatchToProps,
+    )(EditLearningResource),
+  ),
 );
