+++ src/server/contentSecurityPolicy.js
@@ -17,6 +17,7 @@ const connectSrc = (() => {
     'https://hlsak-a.akamaihd.net',
     'https://*.zendesk.com',
     'https://*.boltdns.net',
+    'https://www.google-analytics.com',
     'https://www.googleapis.com/customsearch/',
     'https://house-fastly-signed-eu-west-1-prod.brightcovecdn.com',
   ];
@@ -75,6 +76,10 @@ const scriptSrc = (() => {
     '*.ndla.no',
     'cdnjs.cloudflare.com',
     'https://*.auth0.com',
+    'https://www.googletagmanager.com',
+    'https://www.google-analytics.com',
+    'https://tagmanager.google.com',
+    'http://www.google-analytics.com',
   ];
   if (process.env.NODE_ENV === 'development') {
     return [
@@ -165,6 +170,7 @@ export default {
       'https://i.ytimg.com/',
       'https://ssl.gstatic.com',
       'https://www.gstatic.com',
+      'https://www.google-analytics.com',
       ' data:',
       'blob:',
     ],
